Donot remove virtualenv for RHEL-8
Since RHEL-8 virtualenv rpm works fine and is not incompatible.
Removing venv there breaks the whole RHEL8 jobs. So keep the
workaround for CentOS 7 only.

For CentOS-8, we will be using Python3 so it will not affect.

Closes-Bug: #1862909

+++ install-deps.sh
@@ -180,9 +180,10 @@ install_virtual_env(){
             # TODO(chkumar): virtualenv:20.0.1 is broken due to
             # https://github.com/pypa/virtualenv/issues/1551 and installing lower version
             # will fix the issue. It will be removed in future once it gets fixed.
-            sudo $(package_manager) remove -y $VIRTUALENV_PACKAGE
-            sudo $(python_cmd) -m pip uninstall -y virtualenv
-            sudo $(python_cmd) -m pip install "virtualenv<20.0.0"
+            if [ "$(python_cmd)" != "python3" ]; then
+                sudo $(python_cmd) -m pip uninstall -y virtualenv
+                sudo $(python_cmd) -m pip install "virtualenv<20.0.0"
+            fi
             $(python_cmd) -m virtualenv \
                 $( [ "$OPT_SYSTEM_PACKAGES" = 1 ] && printf -- "--system-site-packages\n" )\
                 $OPT_WORKDIR
